Removes date and datetimes variables

step_yeet creates a specification of a recipe step that will remove date and datetime variables. (This is a meme)

step_yeet(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  removals = NULL,
  skip = FALSE,
  id = rand_id("yeet")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables for this step. See selections() for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

removals
A character string that contains the names of columns that should be removed. These values are not determined until prep.recipe() is called.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Examples

library(recipes)
dat <- data.frame(
  dt = as.POSIXct("2012-01-15"),
  d = as.Date("2020-01-01"),
  n = 4,
  c = "h"
)

recipe(~., data = dat) %>%
  step_yeet() %>%
  prep() %>%
  bake(new_data = NULL)

#> # A tibble: 1 × 2
#>       n c    
#>   <dbl> <fct>
#> 1     4 h